Added json parsing and client to server upload
- Added ActivityJson as json format for json reading
- Refactored wh to long as some activities have a number that is to big
  for int.

No comments or tests yet, will do some code cleanup comming week for
this class.

+++ commons/src/main/java/commons/CompareQuestion.java
@@ -6,7 +6,7 @@ public class CompareQuestion extends Question_Bryan {
      * @param answers The answers for this question, index 0 is the correct one.
      * @param wattage The wattage's for this question, index 0 is the correct one.
      */
-    public CompareQuestion(String[] answers, int[] wattage) {
+    public CompareQuestion(String[] answers, long[] wattage) {
         super(answers, wattage);
     }
 }
